# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de GreenCupones App y movil",
        "description": "Documentación de la API para gestionar productos y ventas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/"
        }
    ],
    "paths": {
...
        "/patients": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Obtener todos los pacientes",
                "operationId": "90275bb01ca424c9f43eba1310bcfc31",
                "responses": {
                    "200": {
                        "description": "Lista de pacientes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Crear un nuevo paciente",
                "operationId": "f339d0077ba97fd7c2dcf90987d9b0cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "last_name",
                                    "type_document",
                                    "number_document",
                                    "date_of_birth",
                                    "email",
                                    "cellphone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "type_document": {
                                        "type": "string"
                                    },
                                    "number_document": {
                                        "type": "string"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "cellphone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paciente creado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/patients/{id}": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Obtener un paciente por ID",
                "operationId": "4731766140224a50c904dbd45837e97f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente encontrado"
                    },
                    "404": {
                        "description": "Paciente no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Actualizar un paciente",
                "operationId": "54fa3d415ed1a308b0cacfbc957869ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "type_document": {
                                        "type": "string"
                                    },
                                    "number_document": {
                                        "type": "string"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "cellphone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paciente actualizado"
                    },
                    "404": {
                        "description": "Paciente no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Eliminar un paciente",
                "operationId": "358e215d5853928883fdd1098b321a4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Paciente eliminado"
                    },
                    "404": {
                        "description": "Paciente no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/patients/{search}/{type}": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Obtener un paciente por ID",
                "operationId": "6da713cf5ee0137ff4d0d6e04af9a1cf",
                "parameters": [
                    {
                        "name": "search",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Si deseas consultar por celular debes ingresar en este campo cellphone, si es cédula dejarlo vacio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente encontrado"
                    },
                    "404": {
                        "description": "Paciente no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/patients/restore/{id}": {
            "put": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Activar un paciente",
                "operationId": "7fa6741b37f607c89c2dae4faa3e8666",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente activo"
                    },
                    "404": {
                        "description": "Paciente no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticación para acceder a la API",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}